[Decomp][Player] Cohesion split: extract Player::LoadFromDB to PlayerLoadFromDB.cpp (#278)
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Ported-from: mangosthree/server@625b038578c2
